Implement V4 with tests. Main monotonic order id for tie-breaking during sort.

+++ include/Trade.h
@@ -2,12 +2,12 @@
 #include "Order.h"
 
 struct Trade {
-    int buyOrderId;
-    int sellOrderId;
+    uint64_t buyOrderId;
+    uint64_t sellOrderId;
     int executionPrice;
     int executionQuantity;
 
-    Trade(int b, int s, int p, int q)
+    Trade(uint64_t b, uint64_t s, int p, int q)
         : buyOrderId(b), sellOrderId(s), executionPrice(p), executionQuantity(q) {}
 };
